+++ web/app/themes/welcome.helsinki/resources/views/blocks/react-share.blade.php
@@ -1,0 +1,1 @@
+@include('partials/react-and-share')