+++ web/app/themes/welcome.helsinki/resources/views/partials/react-and-share.blade.php
@@ -1,0 +1,12 @@
+<div class="rns"></div>
+<script type="text/javascript">
+  (function() {
+    window.rnsData = {
+      apiKey: 'oyp8ikjx4fndltql'
+    };
+    var s = document.createElement('script');
+    s.src = 'https://cdn.reactandshare.com/plugin/rns.js';
+
+    document.body.appendChild(s);
+  }());
+</script>